Move builder/preprocessor into its own file.
